feat: implement javascript types for Value class, fix circular dep
