+++ packages/graph-ts-ponder/src/common/value.ts
@@ -2,7 +2,7 @@ import type { bool, f64, i32, i64, u32, u64 } from '../inject'
 import { assert, changetype } from '../inject'
 import { Bytes, TypedMap } from './collections'
 import { json } from './json'
-import { Address, BigDecimal, BigInt } from './numbers'
+import { Address, BigDecimal } from './numbers'
 
 /**
  * Enum for supported value types.
@@ -33,8 +33,28 @@ const VALUE_KIND_NAMES = [
  * Pointer type for Value data.
  *
  * Big enough to fit any pointer or native `this.data`.
+ *
+ * PONDER: Here are the types by ValueKind:
+ *
+ * STRING: string
+ * INT: number
+ * BIGDECIMAL: BigDecimal
+ * BOOL: boolean
+ * ARRAY: Value[]
+ * NULL: null
+ * BYTES: Bytes
+ * BIGINT: bigint
  */
-export type ValuePayload = u64
+
+export type ValuePayload =
+  | string
+  | number
+  | BigDecimal
+  | boolean
+  | Value[]
+  | null
+  | Bytes
+  | bigint
 
 /**
  * A dynamically typed value.
@@ -44,7 +64,7 @@ export class Value {
 
   toAddress(): Address {
     assert(this.kind == ValueKind.BYTES, 'Value is not an address.')
-    return changetype<Address>(this.data as u32)
+    return Address.fromBytes(this.data as Bytes)
   }
 
   toBoolean(): boolean {
@@ -52,12 +72,12 @@ export class Value {
       return false
     }
     assert(this.kind == ValueKind.BOOL, 'Value is not a boolean.')
-    return this.data != 0
+    return this.data as boolean
   }
 
   toBytes(): Bytes {
     assert(this.kind == ValueKind.BYTES, 'Value is not a byte array.')
-    return changetype<Bytes>(this.data as u32)
+    return this.data as Bytes
   }
 
   toI32(): i32 {
@@ -65,81 +85,54 @@ export class Value {
       return 0
     }
     assert(this.kind == ValueKind.INT, 'Value is not an i32.')
-    return this.data as i32
+    // NOTE: Test this.
+    return this.data as number
   }
 
   toString(): string {
     assert(this.kind == ValueKind.STRING, 'Value is not a string.')
-    return changetype<string>(this.data as u32)
+    return this.data as string
   }
 
   toBigInt(): bigint {
     assert(this.kind == ValueKind.BIGINT, 'Value is not a BigInt.')
-    return changetype<bigint>(this.data as u32)
+    // NOTE: Test this.
+    return this.data as bigint
   }
 
   toBigDecimal(): BigDecimal {
     assert(this.kind == ValueKind.BIGDECIMAL, 'Value is not a BigDecimal.')
-    return changetype<BigDecimal>(this.data as u32)
+    // NOTE: Test this.
+    return new BigDecimal(this.data as bigint)
   }
 
   toArray(): Array<Value> {
     assert(this.kind == ValueKind.ARRAY, 'Value is not an array.')
-    return changetype<Array<Value>>(this.data as u32)
+    return this.data as Value[]
   }
 
   toBooleanArray(): Array<boolean> {
-    const values = this.toArray()
-    const output = new Array<boolean>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBoolean()
-    }
-    return output
+    return this.toArray().map((val) => val.toBoolean())
   }
 
   toBytesArray(): Array<Bytes> {
-    const values = this.toArray()
-    const output = new Array<Bytes>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBytes()
-    }
-    return output
+    return this.toArray().map((val) => val.toBytes())
   }
 
   toStringArray(): Array<string> {
-    const values = this.toArray()
-    const output = new Array<string>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toString()
-    }
-    return output
+    return this.toArray().map((val) => val.toString())
   }
 
   toI32Array(): Array<i32> {
-    const values = this.toArray()
-    const output = new Array<i32>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toI32()
-    }
-    return output
+    return this.toArray().map((val) => val.toI32())
   }
 
   toBigIntArray(): Array<bigint> {
-    const values = this.toArray()
-    const output = new Array<bigint>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBigInt()
-    }
-    return output
+    return this.toArray().map((val) => val.toBigInt())
   }
 
   toBigDecimalArray(): Array<BigDecimal> {
-    const values = this.toArray()
-    const output = new Array<BigDecimal>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBigDecimal()
-    }
-    return output
+    return this.toArray().map((val) => val.toBigDecimal())
   }
 
   /** Return a string that indicates the kind of value `this` contains for
@@ -184,95 +177,67 @@ export class Value {
   }
 
   static fromBooleanArray(input: Array<boolean>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBoolean(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromBoolean(val)))
   }
 
   static fromBytesArray(input: Array<Bytes>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBytes(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromBytes(val)))
   }
 
   static fromI32Array(input: Array<i32>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromI32(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromI32(val)))
   }
 
   static fromBigIntArray(input: Array<bigint>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBigInt(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromBigInt(val)))
   }
 
   static fromBigDecimalArray(input: Array<BigDecimal>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBigDecimal(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromBigDecimal(val)))
   }
 
   static fromStringArray(input: Array<string>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromString(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromString(val)))
   }
 
   static fromAddressArray(input: Array<Address>): Value {
-    const output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromAddress(input[i])
-    }
-    return Value.fromArray(output)
+    return Value.fromArray(input.map((val) => Value.fromAddress(val)))
   }
 
   static fromArray(input: Array<Value>): Value {
-    return new Value(ValueKind.ARRAY, changetype<u32>(input))
+    return new Value(ValueKind.ARRAY, input)
   }
 
   static fromBigInt(n: bigint): Value {
-    return new Value(ValueKind.BIGINT, changetype<u32>(n))
+    return new Value(ValueKind.BIGINT, n)
   }
 
   static fromBoolean(b: bool): Value {
-    return new Value(ValueKind.BOOL, b ? 1 : 0)
+    return new Value(ValueKind.BOOL, b)
   }
 
   static fromBytes(bytes: Bytes): Value {
-    return new Value(ValueKind.BYTES, changetype<u32>(bytes))
+    return new Value(ValueKind.BYTES, bytes)
   }
 
   static fromNull(): Value {
-    return new Value(ValueKind.NULL, 0)
+    return new Value(ValueKind.NULL, null)
   }
 
   static fromI32(n: i32): Value {
-    return new Value(ValueKind.INT, n as u64)
+    return new Value(ValueKind.INT, n)
   }
 
   static fromString(s: string): Value {
-    return new Value(ValueKind.STRING, changetype<u32>(s))
+    return new Value(ValueKind.STRING, s)
   }
 
   static fromAddress(s: Address): Value {
-    return new Value(ValueKind.BYTES, changetype<u32>(s))
+    return new Value(ValueKind.BYTES, s)
   }
 
   static fromBigDecimal(n: BigDecimal): Value {
-    return new Value(ValueKind.BIGDECIMAL, changetype<u32>(n))
+    return new Value(ValueKind.BIGDECIMAL, n)
   }
 }
